Reorganize source code for Elements; obsolete those we don't use
