+++ src/TestCentric/testcentric.gui/Elements/ControlElements/CheckBoxElement.cs
