+++ src/TestCentric/testcentric.gui/Elements/IListBox.cs
@@ -1,8 +1,9 @@
-﻿// ***********************************************************************
+// ***********************************************************************
 // Copyright (c) Charlie Poole and TestCentric contributors.
 // Licensed under the MIT License. See LICENSE file in root directory.
 // ***********************************************************************
 
+using System;
 using System.Windows.Forms;
 
 namespace TestCentric.Gui.Elements
@@ -12,6 +13,7 @@ namespace TestCentric.Gui.Elements
     /// representing a ListBox containing string items or
     /// items that implement ToString() in a useful way.
     /// </summary>
+    [Obsolete("No longer used", true)]
     public interface IListBox : IControlElement
     {
         ListBox.ObjectCollection Items { get; }
