+++ src/TestCentric/testcentric.gui/Elements/ToolStripElements/CheckedMenuElement.cs
