+++ src/TestCentric/testcentric.gui/Elements/ToolStripElements/CheckedToolStripMenuGroup.cs
