+++ src/TestCentric/testcentric.gui/Elements/ToolStripElements/MultiCheckedToolStripButtonGroup.cs
