+++ src/TestCentric/testcentric.gui/Elements/ToolStripElements/ToolStripButtonElement.cs
@@ -8,26 +8,27 @@ using System.Windows.Forms;
 namespace TestCentric.Gui.Elements
 {
     /// <summary>
-    /// ButtonElement wraps a Button as an ICommand.
+    /// ToolStripButtonElement wraps a Windows ToolStripButton.
     /// </summary>
-    public class ButtonElement : ControlElement, ICommand
+    public class ToolStripButtonElement : ToolStripElement, ICommand, IChecked
     {
-        private Button _button;
+        private ToolStripButton _button;
 
         public event CommandHandler Execute;
-        //public event CommandHandler CheckedChanged;
+        public event CommandHandler CheckedChanged;
 
-        //public bool Checked
-        //{
-        //    get { return _button.Checked; }
-        //    set { InvokeIfRequired(() => _button.Checked = value); }
-        //}
+        public bool Checked
+        {
+            get { return _button.Checked; }
+            set { InvokeIfRequired(() => _button.Checked = value); }
+        }
 
-        public ButtonElement(Button button) : base(button)
+        public ToolStripButtonElement(ToolStripButton button) : base(button)
         {
             _button = button;
 
             button.Click += (s, e) => Execute?.Invoke();
+            button.CheckedChanged += (s, e) => CheckedChanged.Invoke();
         }
     }
 }
