+++ src/TestCentric/testcentric.gui/Elements/ToolStripElements/ToolStripElement.cs
